refactor: 상수 클래스 renaming
