+++ src/main/java/lotto/domain/LottoMachine.java
@@ -5,7 +5,7 @@ import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
 import java.util.List;
 
-import static lotto.domain.constant.DomainConstant.*;
+import static lotto.domain.constant.LottoConstant.*;
 
 public class LottoMachine {
