+++ src/main/java/lotto/domain/constant/LottoConstant.java
@@ -1,6 +1,6 @@
 package lotto.domain.constant;
 
-public class DomainConstant {
+public class LottoConstant {
 
     //Money
     public static int MONEY_AMOUNT_UNIT = 1000;
@@ -10,8 +10,6 @@ public class DomainConstant {
     public static int END_LOTTO_NUMBER = 45;
     public static int LOTTO_NUMBER_COUNT = 6;
     public static int FIFTH_PRIZE_COUNT = 3;
-    public static int FOURTH_PRIZE_COUNT = 4;
-    public static int THIRD_PRIZE_COUNT = 5;
     public static String LOTTO_NUMBER_DELIMITER = ", ";
     public static String LOTTO_NUMBER_PREFIX = "[";
     public static String LOTTO_NUMBER_SUFFIX = "]";
